3. Данные и индексы/src/requirements.txt
appnope==0.1.3
Django==3.2
django-filter==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.15.1
flake8==4.0.1
ipython==8.4.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-env==0.6.2
pytest-flake8==1.1.1
pytest-isort==3.0.0
pytest-mock==3.8.2
pytest-pythonpath==0.7.3
pytest-sugar==0.9.5
pytest-test-groups==1.0.3
pytest-testmon==1.3.4
python-dateutil==2.8.2
gunicorn==20.1.0
